Include the Device in the miscdev::Registration
Since the newer version does not compile without it being included i thought i would add this change. Note that this is my first time diving into kernel code so the use of unwrap may not be appropriate so if anyone knows a better solution then please come forth with it.